Rescue corrupted dev branch.
    Set z-offsets for car surfaces.

    Add optional z attribute to MCSurface / surface config loader.

    Fix exception handling.
    - Re-implement QApplication::notify()

    Rename textures.conf => surfaces.conf

    Add mcvector2d.hh and mcvector3d.hh to CMakeLists.txt

    Divide layers into render and collision layers.

    Fix depth buffering.

    Add option to filter layers to be rendered.

    Cosmetics.

    Use C++11 enum class in ResetAction.

    Use C++11 enum class in InputHandler.

    Use C++11 enum class in Game.

    Use C++11 enum class in MTFH::Menu.

    Use C++11 enum class for state.

    Use C++11 enum class for layers.

    Editor: Improve some tile graphics.

    Editor: Different colors for brake and brake hard.

    Refactor the use of gravity.

    Make timing based on frames instead of QTime.

    Fix a bug where mouse release outside pressed item caused two items being focused.

    Instantly change the virtual resolution between fullscreen modes.

+++ src/game/MiniCore/Asset/mcsurfacemetadata.hh
@@ -33,10 +33,10 @@ struct MCSurfaceMetaData
     MCSurfaceMetaData()
     : colorKeySet(false)
     , xAxisMirror(false)
-    , z0(0)
-    , z1(0)
-    , z2(0)
-    , z3(0)
+    , z0(0.0f)
+    , z1(0.0f)
+    , z2(0.0f)
+    , z3(0.0f)
     {}
 
     //! Handle of the surface
@@ -105,16 +105,16 @@ struct MCSurfaceMetaData
     std::pair<GLfloat, bool> specularCoeff;
 
     //! Z-coordinate of vertex[0].
-    int z0;
+    float z0;
 
     //! Z-coordinate of vertex[1].
-    int z1;
+    float z1;
 
     //! Z-coordinate of vertex[2].
-    int z2;
+    float z2;
 
     //! Z-coordinate of vertex[3].
-    int z3;
+    float z3;
 };
 
 #endif // MCSURFACEMETADATA_HH
